Fixed problems with actions
